% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{batch_CalWeight}
\alias{batch_CalWeight}
\title{Title batch calculate weight for a target list and a compendia of comparing lists}
\usage{
batch_CalWeight(targetList, feature.list, method = "Ochiai")
}
\arguments{
\item{targetList}{A vector of gene symbols}

\item{feature.list}{Gene Signatures provided by Molecular Signature Database (MSigDB)}

\item{method}{"Ochiai" or "Jaccard"}
}
\value{
Weight of target genes
}
\description{
Title batch calculate weight for a target list and a compendia of comparing lists
}
\examples{
batch_CalWeight(targetList=targetList,feature.list=feature.list,method="Ochiai")
}
