% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniConSig.R
\name{cal.uniConSig.ks}
\alias{cal.uniConSig.ks}
\title{Title Calculate uniConSig scores based on ks test results}
\usage{
cal.uniConSig.ks(
  up.ks,
  down.ks,
  preCalmatrix,
  feature.list,
  p.cut = 0.01,
  q.cut = 0.25,
  NES.cut = 0,
  power = 1,
  root = 1,
  ECNpenalty = 0.5
)
}
\arguments{
\item{up.ks}{KS test results by weight genes}

\item{down.ks}{KS test reults by -weight genes}

\item{preCalmatrix}{Precalculated matrix to penalize the redundancies in molecular concepts}

\item{feature.list}{pre-compiled molecular concepts}

\item{p.cut}{Default is 0.01}

\item{q.cut}{Default is 0.25}

\item{NES.cut}{Default is 0}

\item{power}{Default is 1}

\item{root}{Default is 1}

\item{ECNpenalty}{Default is 0.5}
}
\value{
uniConSig output. Data.frame of up.uniConSig and down.uniConSig scores for whole genes
}
\description{
Title Calculate uniConSig scores based on ks test results
}
\examples{
cal.uniConSig.ks(up.ks=ks.result[[1]],down.ks=ks.result[[2]],preCalmatrix,feature.list)
}
