% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{cal_ES}
\alias{cal_ES}
\title{Title calculate ES based on gene weight (sorted in decreasing order)}
\usage{
cal_ES(weight, posList)
}
\arguments{
\item{weight}{A vector of gene weight}

\item{posList}{index of genes in the concept list that are common with weighted genes.}
}
\value{
Calculated enrichment score
}
\description{
Title calculate ES based on gene weight (sorted in decreasing order)
}
\examples{
cal_ES(weight, posList=VectorGeneSymbol)
}
