% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{CalWeight}
\alias{CalWeight}
\title{Title Calculate weight based on Ochiai Index}
\usage{
CalWeight(targetList, feature.list, method = "Ochiai")
}
\arguments{
\item{targetList}{Target gene}

\item{feature.list}{Genes symbols in Gene Signatures (MSigDB)}

\item{method}{"Ochiai" or "Jaccard"}
}
\value{
Calculated weight of target genes
}
\description{
Title Calculate weight based on Ochiai Index
}
\examples{
CalWeight(targetList,feature.list,method="Ochiai")
}
