% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ClassData}
\alias{ClassData}
\title{Class data in cls file format}
\format{
Data.frame of class data. The first two rows were removed.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/indepthPathway}
}
\usage{
ClassData
}
\description{
Class data in cls file format
}
\keyword{datasets}
