% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clsFile}
\alias{clsFile}
\title{cls file}
\format{
txt file
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
clsFile
}
\description{
cls file
}
\keyword{datasets}
