% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compare.list}
\alias{compare.list}
\title{Gene signature data obtained from MSigDB}
\format{
A list format. List names are gene signatures and elements are gene symbols
\describe{
\item{KEGG_CITRATE_CYCLE_TCA_CYCLE}{Gene symbols in the concept of KEGG_CITRATE_CYCLE_TCA_CYCLE}
\item{KEGG_PENTOSE_PHOSPHATE_PATHWAY}{Gene symbols in the concept of KEGG_PENTOSE_PHOSPHATE_PATHWAY}
}
}
\source{
\url{https://github.com/wangxlab/indepthPathway}
}
\usage{
compare.list
}
\description{
Gene signature data provided by Molecular Signature Database (MSigDB)
}
\keyword{datasets}
