% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSEA2.R
\name{CSEA2}
\alias{CSEA2}
\title{Title perform CSEA2 for pathway enrichment analysis based on a dichotomous target gene list}
\usage{
CSEA2(
  target.score,
  compare.list,
  p.cut = 0.05,
  minsize = 5,
  min.numOnList = 5,
  transformNegWeight = FALSE,
  cal.qValue = T
)
}
\arguments{
\item{target.score}{Up or Down.uniConSig scores of whole genes.}

\item{compare.list}{Gene Signatures provided by Molecular Signature Database (MSigDB)}

\item{p.cut}{To cut off the pathways by p-value threshold}

\item{minsize}{To cut off gene signatures (MSigDB) by the minimun number of genes}

\item{min.numOnList}{The minimum numbeer of common genes between gene signatures and uniConSig score genes}

\item{transformNegWeight}{To transform uniConSig score or normalize it.}

\item{cal.qValue}{From pathway enrichment}
}
\value{
CSEA or WCSEA output. Data.frame of pathway enrichment.
}
\description{
Title perform CSEA2 for pathway enrichment analysis based on a dichotomous target gene list
}
\examples{
CSEA2(target.score=setNames(as.numeric(uniConSig.result$up.uniConSig), uniConSig.result$subjectID),compare.list,p.cut=1)
}
