% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deambiguation.R
\name{deambiguation.CSEA}
\alias{deambiguation.CSEA}
\title{Title Deambiguate pathway result by CSEA or WCSEA}
\usage{
deambiguation.CSEA(
  GEA.result,
  uniConSig.result,
  compare.list,
  upPathways = NULL,
  topn = 30,
  p.cut = 0.05,
  p.adjust.method = "BH"
)
}
\arguments{
\item{GEA.result}{Pathway enrichment results}

\item{uniConSig.result}{uniConSig result}

\item{compare.list}{Gene Signatures provided by Molecular Signature Database (MSigDB)}

\item{upPathways}{TRUE or FALSE. TRUE will use up.uniConSig scores}

\item{topn}{The number of pathways to deambiguate.}

\item{p.cut}{Default is 0.05}

\item{p.adjust.method}{A method to adjust p-values; "BH","bonferroni" or NULL}
}
\value{
Deambiguated pathway enrichment result
}
\description{
Title Deambiguate pathway result by CSEA or WCSEA
}
\examples{
disambiguation.CSEA(GEA.result=down.CSEA.result,uniConSig.result=uniConSig.result,compare.list=compare.list,upPathways=FALSE,topn=30,p.cut=0.01,p.adjust.method="bonferroni")
}
