% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deambiguation.R
\name{deambiguation.GSEA}
\alias{deambiguation.GSEA}
\title{Title Deambiguate pathway result by GSEA}
\usage{
deambiguation.GSEA(
  GEA.result,
  weight,
  compare.list,
  topn = 30,
  p.cut = 0.05,
  transformNegWeight = FALSE,
  p.adjust.method = "BH"
)
}
\arguments{
\item{GEA.result}{Pathway enrichment results}

\item{weight}{target gene weight}

\item{compare.list}{Gene Signatures provided by Molecular Signature Database (MSigDB)}

\item{topn}{The number of pathways to deambiguate.}

\item{p.cut}{Default is 0.05}

\item{transformNegWeight}{To transform uniConSig score or normalize it.}

\item{p.adjust.method}{A method to adjust p-values; "BH","bonferroni" or NULL}
}
\value{
Deambiguated pathway enrichment result by GSEA
}
\description{
Title Deambiguate pathway result by GSEA
}
\examples{
deambiguation.GSEA(GEA.result=down.GSEA.result,weight=-weight,compare.list=compare.list,topn=min(c(topn,nrow(down.GSEA.result))),p.cut=0.8,p.adjust.method="bonferroni")
}
