% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{disambiguationCSEA}
\alias{disambiguationCSEA}
\title{Title disambiguate pathway result by CSEA or WCSEA}
\usage{
disambiguationCSEA(
  GEA.result,
  uniConSigResult,
  compare.list,
  upPathways = NULL,
  topn = 30,
  p.cut = 0.05,
  p.adjust.method = "BH"
)
}
\arguments{
\item{GEA.result}{Pathway enrichment results}

\item{uniConSigResult}{uniConSig result}

\item{compare.list}{Gene Signatures provided by Molecular Signature Database (MSigDB)}

\item{upPathways}{TRUE or FALSE. TRUE will use up.uniConSig scores}

\item{topn}{The number of pathways to disambiguate.}

\item{p.cut}{Default is 0.05}

\item{p.adjust.method}{A method to adjust p-values; "BH","bonferroni" or NULL}
}
\value{
disambiguated pathway enrichment result
}
\description{
Title disambiguate pathway result by CSEA or WCSEA
}
\examples{
disambiguationCSEA(GEA.result=down.CSEA.result,uniConSigResult=uniConSigResult,compare.list=compare.list,upPathways=FALSE,topn=30,p.cut=0.01,p.adjust.method="bonferroni")
}
