% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{down.assoc}
\alias{down.assoc}
\title{Calculate associations between top pathways.}
\format{
vector
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
down.assoc
}
\description{
Calculate associations between selected top down-regulated pathways.
}
\keyword{datasets}
