% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{down.CSEA.result}
\alias{down.CSEA.result}
\title{down.CSEA.result example}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
down.CSEA.result
}
\description{
down.CSEA.result is obtained by CSEA2() when you input down.uniConSig scores
}
\keyword{datasets}
