% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{down.disambiguate}
\alias{down.disambiguate}
\title{Disambiguation of CSEA2 result}
\format{
vector
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
down.disambiguate
}
\description{
Disambiguate downregulated pathways
}
\keyword{datasets}
