% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{down.GSEA.result}
\alias{down.GSEA.result}
\title{down.CSEA.result example}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
down.GSEA.result
}
\description{
down.GSEA.result is obtained by CSEA2() when input gene weight negative
}
\keyword{datasets}
