% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{downpathway}
\alias{downpathway}
\title{Top 30 pathways from down.disambituated pathways.}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
downpathway
}
\description{
Top 30 down.pathways.
}
\keyword{datasets}
