% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{draw.interactome.network}
\alias{draw.interactome.network}
\title{Title Draw interactome network for a selected pathway based on DEG weight}
\usage{
draw.interactome.network(
  links,
  genelist,
  geneweight,
  target = NULL,
  pdf.file,
  self.interaction = F,
  filter.coreEnrich = T
)
}
\arguments{
\item{links}{Interactome file, for example "interactome_all.hs.sym"}

\item{genelist}{Gene symbols in a specific gene signature}

\item{geneweight}{Weight gene}

\item{target}{Default is NULL}

\item{pdf.file}{Output file name.}

\item{self.interaction}{Default is FALSE}

\item{filter.coreEnrich}{Default is TRUE}
}
\value{
Interactome network for a selected pathway based on DEG weight
}
\description{
Title Draw interactome network for a selected pathway based on DEG weight
}
\examples{
draw.interactome.network(links=HuRI.interactions,genelist=genelist,geneweight=geneweight,target=NULL,pdf.file="My.HuRI.network.pdf",self.interaction=F,filter.coreEnrich=T)
}
