% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{draw.network}
\alias{draw.network}
\title{Title Option2.Draw network figure for significant pathways based on association file}
\usage{
draw.network(
  pathway.out,
  assoc,
  NES.cut = 2,
  node.size = 2,
  line.thickness = 1,
  font.size = 0.3,
  wrap.string.width = 10,
  FirstCharacterUpperCase = F
)
}
\arguments{
\item{pathway.out}{Merged pathway data}

\item{assoc}{pathway.merge.assoc}

\item{NES.cut}{Default is 2}

\item{node.size}{Default is 2}

\item{line.thickness}{Default is 1}

\item{font.size}{Default is 0.3}

\item{wrap.string.width}{Default is 10}

\item{FirstCharacterUpperCase}{Default is FALSE}
}
\value{
network of the functional associations between selected top pathways
}
\description{
Title Option2.Draw network figure for significant pathways based on association file
}
\examples{
draw.network(pathway.out=pathway.merge,assoc=pathway.merge.assoc,NES.cut=2,node.size=1,line.thickness=1,font.size=0.4,wrap.string.width=15)
}
