% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SymSim.R
\name{expand2binary}
\alias{expand2binary}
\title{Title Expand transcript counts to a vector of binaries of the same length of as the number of transcripts}
\usage{
expand2binary(true_counts_1cell)
}
\arguments{
\item{true_counts_1cell}{number of transcript in one cell}
}
\value{
list of expanded binary data.
}
\description{
Title Expand transcript counts to a vector of binaries of the same length of as the number of transcripts
}
\examples{
expand2binary(true_counts_1cell)
}
