% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExpData}
\alias{ExpData}
\title{Gene expression data in gct file format}
\format{
Data.frame of gene expression data. The first two rows were removed.
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
ExpData
}
\description{
Gene expression data in gct file format
}
\keyword{datasets}
