% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExpMatrix}
\alias{ExpMatrix}
\title{Expression count matrix example. You can make GCT from this matrix.}
\format{
matrix
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
ExpMatrix
}
\description{
Expression count matrix  obtained from True2ObservedCounts.
}
\keyword{datasets}
