% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExpressionData}
\alias{ExpressionData}
\title{Normalized ExpressionData of scRNA-seq.}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
ExpressionData
}
\description{
ExpressionData can be obtained from gct file.
}
\keyword{datasets}
