% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature.list}
\alias{feature.list}
\title{Pre-compiled molecular concepts}
\format{
A list format. List names are concepts and elements are gene symbols
\describe{
\item{KEGG_CITRATE_CYCLE_TCA_CYCLE}{Gene symbols in the concept of KEGG_CITRATE_CYCLE_TCA_CYCLE}
\item{KEGG_PENTOSE_PHOSPHATE_PATHWAY}{Gene symbols in the concept of KEGG_PENTOSE_PHOSPHATE_PATHWAY}
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
feature.list
}
\description{
A data set of pre-compiled molecular concepts
}
\keyword{datasets}
