% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedKSTest.R
\name{filterGeneSymb}
\alias{filterGeneSymb}
\title{Title Filter out gene symbols that don't exist in the DEG (weight) gene symbols}
\usage{
filterGeneSymb(x, WeightGeneSymbol, inlist)
}
\arguments{
\item{x}{pre-compiled molecular concepts. Each element in the molecular concepot list has a couple of dozens genes.}

\item{WeightGeneSymbol}{Gene symbols of weight genes}

\item{inlist}{Default is TRUE.}
}
\value{
Moleculare concept lists that contains gene symbols common with weight genes
}
\description{
Title Filter out gene symbols that don't exist in the DEG (weight) gene symbols
}
\examples{
filterGeneSymb(x=feature.list, WeightGeneSymbol=names(weight),inlist=TRUE)
}
