% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gctFile}
\alias{gctFile}
\title{gct file}
\format{
txt file
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
gctFile
}
\description{
gct file
}
\keyword{datasets}
