% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_len}
\alias{gene_len}
\title{The length data for the genes in the ExpressionData}
\format{
a vector
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
gene_len
}
\description{
This gene length data is from SymSim
}
\keyword{datasets}
