% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GenelengthData}
\alias{GenelengthData}
\title{Gene symbol and gene length data}
\format{
Data.frame of gene length data
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
GenelengthData
}
\description{
Gene length data for 11,506 genes
}
\keyword{datasets}
