% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genelist}
\alias{genelist}
\title{genelist example to draw interactome network for a selected pathway and your target gene of interests.}
\format{
vector
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
genelist
}
\description{
You can analyze the interactions of your gene of interests with the pathways genes.
}
\keyword{datasets}
