% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeneSigFile}
\alias{GeneSigFile}
\title{Molecular pathway or gene signature file}
\format{
file name
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
GeneSigFile
}
\description{
Molecular pathway or gene signature file to make compare.list
}
\keyword{datasets}
