% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HuRI.interactions}
\alias{HuRI.interactions}
\title{Gene interaction pair data.}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
HuRI.interactions
}
\description{
This gene data provide information of gene interactions to make a network.
}
\keyword{datasets}
