% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathway-package.R
\docType{package}
\name{IndepthPathway-package}
\alias{IndepthPathway}
\alias{IndepthPathway-package}
\title{IndepthPathway: Identify enriched pathways from gene expression data.}
\description{
IndepthPathway can calculate unoConSig scores and run CSEA, WCSEA, or GSEA to identify enriched pathways.
}
\author{
\strong{Maintainer}: Xiaosong Wang \email{xiaosongw@pitt.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
