% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ks.result}
\alias{ks.result}
\title{KS test to calculate enrichment of feature list}
\format{
list
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
ks.result
}
\description{
The weight of genes are used to compute functional relations of feature list
}
\keyword{datasets}
