% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{len2nfrag}
\alias{len2nfrag}
\title{Transcript length to number of fragments (for the nonUMI protocol)}
\format{
a vector
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
len2nfrag
}
\description{
From transcript length to number of fragments (for the nonUMI protocol)
}
\keyword{datasets}
