% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{limmaDEG}
\alias{limmaDEG}
\title{Limma to find differentially expressed genes}
\usage{
limmaDEG(gctFile, clsFile, groups.order = c("ClassA", "ClassB"))
}
\arguments{
\item{gctFile}{Gene expression data in GCT file format}

\item{clsFile}{Class data in cls format.}

\item{groups.order}{two classes}
}
\value{
Limma input data in list
}
\description{
Limma to find differentially expressed genes
}
\examples{
limmaDEG(gctFile,clsFile,groups.order=c("ClassA","ClassB"))
}
