% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LimmaFit}
\alias{LimmaFit}
\title{LimmaFit data}
\format{
limma
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
LimmaFit
}
\description{
LimmaFit data
}
\keyword{datasets}
