% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LimmaOut}
\alias{LimmaOut}
\title{LimmaOut}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
LimmaOut
}
\description{
Limma Output
}
\keyword{datasets}
