% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{LimmaWeight}
\alias{LimmaWeight}
\title{Limma to find differentially expressed genes}
\usage{
LimmaWeight(LimmaFit = LimmaFit)
}
\arguments{
\item{LimmaFit}{Gene expression data in gct format. But first two rows were removed.}
}
\value{
data.frame of Limma output
}
\description{
Limma to find differentially expressed genes
}
\examples{
LimmaWeight(LimmaFit)
}
