% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{LoadPackage}
\alias{LoadPackage}
\title{Install and load necessary packages}
\usage{
LoadPackage()
}
\value{
install and load packages
}
\description{
Install and load necessary packages
}
\examples{
LoadPackage()
}
