% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SymSim.R
\name{MakeGCT}
\alias{MakeGCT}
\title{Make a gct file format from expression data matrix}
\usage{
MakeGCT(ExpMatrix)
}
\arguments{
\item{ExpMatrix}{Gene expression data matrix. Row names are gene symbols and column names are sample names.}
}
\value{
GCT file format of gene expression matrix
}
\description{
Make a gct file format from expression data matrix
}
\examples{
MakeGCT(ExpMatrix)
}
