% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{merge.pathway}
\alias{merge.pathway}
\title{Title Merge upregulated and downregulated pathways. For pathways that are both up or down regulated, the most significant rank will be retained}
\usage{
\method{merge}{pathway}(up.pathway, down.pathway)
}
\arguments{
\item{up.pathway}{filtered up.deambiguate, up.deambiguate[\link{1}]}

\item{down.pathway}{filtered down.deambiguate, down.deambiguate[\link{1}]}
}
\value{
Merged pathways. The pathways of the most significant rank will be retained
}
\description{
Title Merge upregulated and downregulated pathways. For pathways that are both up or down regulated, the most significant rank will be retained
}
\examples{
merge.pathway(up.pathway=up.deambiguate[[1]][1:30,],down.pathway=down.deambiguate[[1]][1:30,])
}
