% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{mergePathway}
\alias{mergePathway}
\title{Title Merge upregulated and downregulated pathways. For pathways that are both up or down regulated, the most significant rank will be retained}
\usage{
mergePathway(uppathway, downpathway)
}
\arguments{
\item{uppathway}{filtered up_disambiguate}

\item{downpathway}{filtered down_disambiguate}
}
\value{
Merged pathways. The pathways of the most significant rank will be retained
}
\description{
Title Merge upregulated and downregulated pathways. For pathways that are both up or down regulated, the most significant rank will be retained
}
\examples{
mergePathway(uppathway,downpathway)
}
