% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myPermu}
\alias{myPermu}
\title{Permuted calculation of enrichment scores}
\format{
List
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
myPermu
}
\description{
Permuted calculation of enrichment scores
}
\keyword{datasets}
