% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedKSTest.R
\name{normalize}
\alias{normalize}
\title{Title normalize the weight of genes}
\usage{
normalize(x)
}
\arguments{
\item{x}{A vector of gene weight}
}
\value{
normalized weight
}
\description{
Title normalize the weight of genes
}
\examples{
normalize(weight)
}
