% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{normalizeWeight}
\alias{normalizeWeight}
\title{Title normalize the weight of genes}
\usage{
normalizeWeight(x)
}
\arguments{
\item{x}{A vector of gene weight}
}
\value{
normalized weight
}
\description{
Title normalize the weight of genes
}
\examples{
normalize(weight)
}
