% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{pathway.heatmap}
\alias{pathway.heatmap}
\title{Title Option1.Draw heatmaps showing the functional associations between selected top pathways}
\usage{
pathway.heatmap(
  matrixData,
  clustering = TRUE,
  distanceMatric = "euclidean",
  fontSize = 0.7,
  sepWidth = 0.1
)
}
\arguments{
\item{matrixData}{PathwayAssociation data}

\item{clustering}{Default is TRUE}

\item{distanceMatric}{Default is 'euclidean'}

\item{fontSize}{Default is 0.7}

\item{sepWidth}{Default is 0.1}
}
\value{
heatmaps showing the functional associations between selected top pathways
}
\description{
Title Option1.Draw heatmaps showing the functional associations between selected top pathways
}
\examples{
pathway.heatmap(matrixData=up.assoc,clustering = T,fontSize=5)
}
