% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathway.merge}
\alias{pathway.merge}
\title{Merged pathways of top up- and down-regulated}
\format{
vector
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
pathway.merge
}
\description{
Merged pathways of top up- and down-regulated
}
\keyword{datasets}
