% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathwayMergeAssoc}
\alias{pathwayMergeAssoc}
\title{Calculate associations between merged pathways of up- and down-}
\format{
vector
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
pathwayMergeAssoc
}
\description{
Calculate associations between merged up- and down-regulated pathways.
}
\keyword{datasets}
