% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedKSTest.R
\name{perm_weightedKS}
\alias{perm_weightedKS}
\title{Title calculate permutated ES}
\usage{
perm_weightedKS(weight, numOnList, nPermu = 1000)
}
\arguments{
\item{weight}{A vector of gene weight}

\item{numOnList}{index of genes in the concept list that are common with weighted genes.}

\item{nPermu}{Default is 1000}
}
\value{
Permuted enrichment score of weight genes
}
\description{
Title calculate permutated ES
}
\examples{
perm_weightedKS(weight, numOnList=10, nPermu=1000)
}
