% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plot.geneset.matrix}
\alias{plot.geneset.matrix}
\title{Title Option4.Draw heatmap for selected pathway}
\usage{
\method{plot}{geneset.matrix}(
  genomatrix,
  geneset,
  weights,
  fontsize_row = 6,
  fontsize_col = 6,
  cluster_rows = F,
  annCol = NULL
)
}
\arguments{
\item{genomatrix}{Gene expression data}

\item{geneset}{A specific gene signature}

\item{weights}{weight gene}

\item{fontsize_row}{Default is 6}

\item{fontsize_col}{Default is 6}

\item{cluster_rows}{Default is FALSE}

\item{annCol}{Default is NULL}
}
\value{
heatmap of gnes for selected gene signature
}
\description{
Title Option4.Draw heatmap for selected pathway
}
\examples{
plot.geneset.matrix(genomatrix=expData,geneset=compare.list[["HALLMARK_E2F_TARGETS"]],weights=weight,annCol = sampleClass)
}
