% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{PlotGenesetMatrix}
\alias{PlotGenesetMatrix}
\title{Title}
\usage{
PlotGenesetMatrix(
  genomatrix,
  geneset,
  weights,
  fontsize_row = 6,
  fontsize_col = 6,
  cluster_rows = F,
  annCol = NULL,
  filter.coreEnrich = T
)
}
\arguments{
\item{genomatrix}{Gene expression data}

\item{geneset}{A specific gene signature}

\item{weights}{weight gene}

\item{fontsize_row}{Default is 6}

\item{fontsize_col}{Default is 6}

\item{cluster_rows}{Default is FALSE}

\item{annCol}{Default is NULL}

\item{filter.coreEnrich}{Default is TRUE}
}
\value{
It generates a heatmap of selected gene signature
}
\description{
Title
}
\examples{
PlotGenesetMatrix(genomatrix=expData,geneset=compare.list[["HALLMARK_E2F_TARGETS"]],weights=weight,annCol=sampleClass)
}
