% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{preCalmatrix}
\alias{preCalmatrix}
\title{Precalculated matrix to penalize the redundancies in molecular concepts}
\format{
A gmt file format.
\describe{
\item{ACSS2}{KEGG_GLYCOLYSIS_GLUCONEOGENESIS@6.40917382433143\tKEGG_PYRUVATE_METABOLISM@6.01715179470204}
\item{GCK}{KEGG_GLYCOLYSIS_GLUCONEOGENESIS@16.9306840571511\tKEGG_GALACTOSE_METABOLISM@13.6040224314813}
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
preCalmatrix
}
\description{
Precalculated matrix to penalize the redundancies in molecular concepts
}
\keyword{datasets}
