% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_concepts.R
\name{read.gct}
\alias{read.gct}
\title{Title Read gct and cls files}
\usage{
read.gct(gctFile, clsFile)
}
\arguments{
\item{gctFile}{my .gct file name}

\item{clsFile}{my .cls file name}
}
\value{
List of expression data and class data
}
\description{
Title Read gct and cls files
}
\examples{
read.gct(gctFile="scData_quiescentVsActive.gct",clsFile="scData_quiescentVsActive.cls")
}
