% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_concepts.R
\name{read_concepts}
\alias{read_concepts}
\title{Read molecular concepts or gene signature from gmt files}
\usage{
read_concepts(fileName, min = 10)
}
\arguments{
\item{fileName}{molecular concept file in gmt format}

\item{min}{to filter out concepts that have genes less than 'min' value}
}
\value{
molecular concepts in list
}
\description{
Read molecular concepts or gene signature from gmt files
}
\examples{
read_concepts("h.all.v7.5.1.symbols.gmt")
}
