% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SymSim.R
\name{rnorm_trunc}
\alias{rnorm_trunc}
\title{sample from truncated normal distribution}
\usage{
rnorm_trunc(n, mean, sd, a, b)
}
\arguments{
\item{a}{the minimum value allowed}

\item{b}{the maximum value allowed}
}
\value{
vector of turncated normal distribution
}
\description{
sample from truncated normal distribution
}
\examples{
rnorm_trunc(n, mean, sd, a, b)
}
