% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{run.weightedKS}
\alias{run.weightedKS}
\title{Title run weighted KS for a weighted subject list using weightedKSV2 module}
\usage{
run.weightedKS(
  weight,
  signed = TRUE,
  feature.list,
  minsize = 10,
  correct.outlier = FALSE,
  transformNegWeight = FALSE
)
}
\arguments{
\item{weight}{signed q-values calculated from limmaDEG or scdeDEG}

\item{signed}{TRUE is default. This means that you will use signed q-value to run weighted k-s test.}

\item{feature.list}{pre-compiled molecular concepts}

\item{minsize}{Filter out concepts that have genes less than minsize}

\item{correct.outlier}{Default is FALSE. User can choose to correct outliers}

\item{transformNegWeight}{Default is FALSE. It transforms the negative tail of the weight into small positive values}
}
\value{
K-S test result.
}
\description{
Title run weighted KS for a weighted subject list using weightedKSV2 module
}
\examples{
run.weightedKS(weight,signed=TRUE,feature.list=feature.list,minsize=10,correct.outlier=FALSE,transformNegWeight=FALSE )
}
