% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{scdeDEG}
\alias{scdeDEG}
\title{SCDE to find differentially expressed genes}
\usage{
scdeDEG(
  ExpDataGCT,
  ClassData,
  groups.order = NULL,
  min.lib.size = 100,
  min.reads = 1,
  min.detected = 1,
  SignPzscore = 0.75
)
}
\arguments{
\item{ExpDataGCT}{Gene expression data in gct format. But first two rows were removed.}

\item{ClassData}{Class data in cls format. But first two rows were removed.}

\item{groups.order}{two classes}

\item{min.lib.size}{default 100}

\item{min.reads}{default 1}

\item{min.detected}{default 1}

\item{SignPzscore}{default 0.75}
}
\value{
data.frame of SCDE output
}
\description{
SCDE to find differentially expressed genes
}
\examples{
scdeDEG(ExpDataGCT, ClassData, groups.order=c("ClassA","ClassB"))
}
