% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targetList}
\alias{targetList}
\title{targetList example}
\format{
a vector
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
targetList
}
\description{
targetList obtained by Limma or SCDE.
}
\keyword{datasets}
