% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targetList_Up}
\alias{targetList_Up}
\title{Genes of Signed.Q.Value > 0 from Limma or SCDE out}
\format{
vector
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
targetList_Up
}
\description{
Limma or SCDE can calculate the weight of genes. The genes of positive weight will be used to calculate uniConSig score
}
\keyword{datasets}
