% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targetScore}
\alias{targetScore}
\title{targetScore example}
\format{
a vector
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
targetScore
}
\description{
targetScore obtained by uniConSig scores for all genes
}
\keyword{datasets}
