% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmp.weight}
\alias{tmp.weight}
\title{tmp.weight example}
\format{
a vector
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
tmp.weight
}
\description{
tmp.weight is the same as targetScore which is obtained by uniConSig scores for all genes
}
\keyword{datasets}
