% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{topPathwayDw}
\alias{topPathwayDw}
\title{Top pathway names from disambiguated CSEA2 out}
\format{
a vector
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
topPathwayDw
}
\description{
Top pathway names of down-regulated.
}
\keyword{datasets}
