% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{true_counts}
\alias{true_counts}
\title{Processed true_counts ExpData of scRNA-seq.}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
true_counts
}
\description{
ExpressionData can be obtained from gct file. This is before simulation.
}
\keyword{datasets}
