% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniConSigResult}
\alias{uniConSigResult}
\title{uniConSigResult example}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
uniConSigResult
}
\description{
Calculate uniConSig scores for all genes
}
\keyword{datasets}
