% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{up.assoc}
\alias{up.assoc}
\title{Calculate associations between top pathways.}
\format{
vector
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
up.assoc
}
\description{
Calculate associations between selected top up-regulated pathways.
}
\keyword{datasets}
