% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{up.CSEA.result}
\alias{up.CSEA.result}
\title{up.CSEA.result example}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
up.CSEA.result
}
\description{
up.CSEA.result is obtained by CSEA2() when you input up.uniConSig scores
}
\keyword{datasets}
