% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{up.GSEA.result}
\alias{up.GSEA.result}
\title{up.GSEA.result example}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
up.GSEA.result
}
\description{
up.GSEA.result is obtained by CSEA2() when you input gene weight positive
}
\keyword{datasets}
