% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uppathway}
\alias{uppathway}
\title{Top 30 pathways from up.disambituated pathways.}
\format{
data.frame
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
uppathway
}
\description{
Top 30 up.pathways.
}
\keyword{datasets}
