% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weight}
\alias{weight}
\title{Weight of genes: Signed.Q.Value}
\format{
vector
\describe{
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
weight
}
\description{
Limma or SCDE can calculate the weight of genes.
}
\keyword{datasets}
