% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{weight.transform}
\alias{weight.transform}
\title{Title transform the negative tail of the weight into small positive values}
\usage{
weight.transform(weight)
}
\arguments{
\item{weight}{signed q-values calculated from limmaDEG or scdeDEG}
}
\value{
positive or negative weight.
}
\description{
Title transform the negative tail of the weight into small positive values
}
\examples{
weight.transform(weight)
weight.transform(-weight)
}
