% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{weightedKS.batch}
\alias{weightedKS.batch}
\title{Title calculated weighted KS using weightedKSV2 based on a gene list with t-values as weight (such as the gene list sorted by t-values)}
\usage{
weightedKS.batch(
  weight,
  feature.list,
  high.enrich = TRUE,
  p.cut = 0.05,
  correct.outlier = FALSE,
  minsize = 5
)
}
\arguments{
\item{weight}{signed q-values of DEG genes calculated from limmaDEG or scdeDEG}

\item{feature.list}{pre-compiled molecular concepts}

\item{high.enrich}{default is TRUE. Then, it uses 'weight' to rank subjects}

\item{p.cut}{Default 0.5}

\item{correct.outlier}{Default is FALSE. User can choose to correct outliers}

\item{minsize}{Filter out concepts that have genes less than minsize}
}
\value{
Data.frame of K-S test results. This K-S test calculate the enrichment of molecular concepts in feature.list
}
\description{
Title calculated weighted KS using weightedKSV2 based on a gene list with t-values as weight (such as the gene list sorted by t-values)
}
\examples{
weightedKS.batch(weight, feature.list=feature.list,high.enrich=TRUE,p.cut=0.05,correct.outlier=FALSE,minsize=5)
}
