% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{weightedKSV2}
\alias{weightedKSV2}
\title{Title Run K-S test for mcolecular concepts}
\usage{
weightedKSV2(
  weight,
  compare.list,
  myPermu,
  p.cut = 0.05,
  min.numOnList = 5,
  cal.qValue = T
)
}
\arguments{
\item{weight}{signed q-values calculated from limmaDEG or scdeDEG}

\item{compare.list}{pre-compiled molecular concepts}

\item{myPermu}{Permuted calculation of enrichment scores}

\item{p.cut}{To cut off the pathways by p-value threshold}

\item{min.numOnList}{The minimum numbeer of common genes between gene signatures and uniConSig score genes}

\item{cal.qValue}{From pathway enrichment}
}
\value{
Weighted K-S test output for molecular concepts.
}
\description{
Title Run K-S test for mcolecular concepts
}
\examples{
weightedKSV2(weight,compare.list,myPermu,p.cut=0.05,min.numOnList=5,cal.qValue=T)
}
