% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{write.cls}
\alias{write.cls}
\title{Title}
\usage{
write.cls(group, cls.file)
}
\arguments{
\item{group}{class of expression data}

\item{cls.file}{my .cls file name}
}
\value{
It writes a .cls file
}
\description{
Title
}
\examples{
write.cls(group, cls.file)
}
