% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{write.gct}
\alias{write.gct}
\title{Title}
\usage{
write.gct(expData, gct.file)
}
\arguments{
\item{expData}{My expression data}

\item{gct.file}{My .gct file}
}
\value{
It writes a .gct file.
}
\description{
Title
}
\examples{
write.gct(expData, gct.file)
}
