% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniConSig.R
\name{cal.uniConSig}
\alias{cal.uniConSig}
\title{Title Calculate uniConSig scores for genes.}
\usage{
cal.uniConSig(
  target.list,
  feature.list,
  preCalmatrix,
  minsize = 10,
  weight.cut = 0.05,
  power = 1,
  root = 1,
  ECNpenalty = 0.5,
  method = "Ochiai",
  rm.overfit
)
}
\arguments{
\item{target.list}{A vector of gene symbols}

\item{feature.list}{pre-compiled molecular concepts}

\item{preCalmatrix}{Precalculated matrix to penalize the redundancies in molecular concepts}

\item{minsize}{Default 10}

\item{weight.cut}{Default 0.05}

\item{power}{Default 1}

\item{root}{Default 1}

\item{ECNpenalty}{Default 0.5}

\item{method}{Default 'Ochiai}

\item{rm.overfit}{Remove overfitted weight genes}
}
\value{
uniConSig output by non-weighted method
}
\description{
Title Calculate uniConSig scores for genes.
}
\examples{
cal.uniConSig(target.list=target.list_Up,feature.list=feature.list,preCalmatrix,rm.overfit=F)
}
