% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathway-package.R
\docType{package}
\name{IndepthPathway-package}
\alias{IndepthPathway}
\alias{IndepthPathway-package}
\title{IndepthPathway: Identify enriched pathways from gene expression data.}
\description{
IndepthPathway can calculate unoConSig scores and run CSEA, WCSEA, or GSEA to identify enriched pathways.
}
\author{
\strong{Maintainer}: Sanghoon Lee \email{sal170@pitt.edu} (\href{https://orcid.org/0000-0001-8564-2392}{ORCID})

}
\keyword{internal}
