% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpressionGene.R
\name{limmaDEG}
\alias{limmaDEG}
\title{Limma to find differentially expressed genes}
\usage{
limmaDEG(ExpDataGCT, ClassData, groups.order = NULL)
}
\arguments{
\item{ExpDataGCT}{Gene expression data in gct format. But first two rows were removed.}

\item{ClassData}{Class data in cls format. But first two rows were removed.}

\item{groups.order}{two classes}
}
\value{
data.frame of Limma output
}
\description{
Limma to find differentially expressed genes
}
\examples{
limmaDEG(ExpDataGCT,ClassData,groups.order=c("ClassA","ClassB"))
}
