% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{LoadPackage}
\alias{LoadPackage}
\title{Install and load necessary packages for IndepthPathway}
\usage{
LoadPackage()
}
\value{
install and load packages for IndepthPathway
}
\description{
Install and load necessary packages for IndepthPathway
}
\examples{
LoadPackage()
}
