% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepthPathwayModules.R
\name{pathwayAssociation}
\alias{pathwayAssociation}
\title{Title Calculation the functional associations of disambiguated pathway enrichment result}
\usage{
pathwayAssociation(
  topPathway = topPathwayUp,
  compare.list,
  feature.list,
  preCalmatrix,
  minsize = 10,
  rm.overfit = FALSE
)
}
\arguments{
\item{topPathway}{up or down.disambiguation data}

\item{compare.list}{Gene Signatures provided by Molecular Signature Database (MSigDB)}

\item{feature.list}{pre-compiled molecular concepts}

\item{preCalmatrix}{Precalculated matrix to penalize the redundancies in molecular concepts}

\item{minsize}{Default 10. Filter out concepts that have genes less than minsize}

\item{rm.overfit}{Default FALSE.}
}
\value{
A similarity matrix between top N pathways
}
\description{
Title Calculation the functional associations of disambiguated pathway enrichment result
}
\examples{
pathwayAssociation(topPathway=topPathwayUp, compare.list, feature.list, preCalmatrix,  minsize=10, rm.overfit=FALSE)
}
