% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{preCalmatrix}
\alias{preCalmatrix}
\title{Precalculated matrix to penalize the redundancies in molecular concepts}
\format{
A gmt file format.
\describe{
\item{ACSS2}{KEGG_GLYCOLYSIS_GLUCONEOGENESIS}{KEGG_PYRUVATE_METABOLISM}
\item{GCK}{KEGG_GLYCOLYSIS_GLUCONEOGENESIS}{KEGG_GALACTOSE_METABOLISM}
}
}
\source{
\url{https://github.com/wangxlab/IndepthPathway}
}
\usage{
preCalmatrix
}
\description{
Precalculated matrix to penalize the redundancies in molecular concepts
}
\keyword{datasets}
